Fixes as per PR response
